change type of smLength in SmppSubmitSmBody from byte to short, in order to support message length 0-254

+++ netty-codec-smpp/src/main/java/io/github/protocol/codec/smpp/SmppEncoder.java
@@ -221,7 +221,7 @@ public class SmppEncoder extends MessageToMessageEncoder<SmppMessage> {
         buf.writeByte(body.replaceIfPresentFlag());
         buf.writeByte(body.dataCoding());
         buf.writeByte(body.smDefaultMsgId());
-        buf.writeByte(body.smLength());
+        buf.writeShort(body.smLength());
         buf.writeBytes(body.shortMessage());
         return buf;
     }
